backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
pdfplumber==0.11.0
PyMuPDF>=1.25.0
google-generativeai==0.5.4
groq==0.9.0
scipy>=1.13.0
pandas>=2.2.0
numpy>=1.26.4
reportlab>=4.1.0
python-dotenv==1.0.1
httpx==0.27.0
pillow>=10.3.0
